requirements.txt.txt
opencv-python
python-chess
stockfish
pandas
numpy
matplotlib
pyautogui
scipy

